MBS-12332 signer plugin rework
register tasks only for configured build variants
