+++ subprojects/delivery/signer/src/main/kotlin/com/avito/plugin/SignBundleTask.kt
@@ -1,6 +1,7 @@
 package com.avito.plugin
 
 import org.gradle.api.file.RegularFile
+import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.Property
 import org.gradle.api.tasks.InputFile
 import org.gradle.workers.WorkerExecutor
@@ -8,11 +9,12 @@ import java.io.File
 import javax.inject.Inject
 
 public abstract class SignBundleTask @Inject constructor(
-    workerExecutor: WorkerExecutor
+    workerExecutor: WorkerExecutor,
+    objects: ObjectFactory,
 ) : AbstractSignTask(workerExecutor) {
 
     @get:InputFile
-    public abstract val bundleFile: Property<RegularFile>
+    public val bundleFile: Property<RegularFile> = objects.fileProperty()
 
     override fun unsignedFile(): File = bundleFile.get().asFile
 }
