+++ subprojects/delivery/signer/src/test/kotlin/com/avito/plugin/UrlResolverTest.kt
@@ -1,5 +1,6 @@
 package com.avito.plugin
 
+import com.avito.plugin.internal.validateUrl
 import com.google.common.truth.Truth.assertThat
 import org.junit.jupiter.api.Test
